rename also kernel filename
