+++ include/dlaf/communication/kernels.h
@@ -13,5 +13,5 @@
 #include "dlaf/communication/kernels/all_reduce.h"
 #include "dlaf/communication/kernels/broadcast.h"
 #include "dlaf/communication/kernels/p2p.h"
-#include "dlaf/communication/kernels/p2p_allreduce.h"
+#include "dlaf/communication/kernels/p2p_allsum.h"
 #include "dlaf/communication/kernels/reduce.h"
